requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.3.1
jupyter_core==5.4.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
stack-data==0.6.3
tiktoken==0.5.1
tokenizers==0.14.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
yarl==1.9.2
zipp==3.17.0
